You now get logged in, as soon as the Icon loads, if there is a cookie
